diff assembly integration tests
